Renamed Game package and moved everything into it.
We needed to rename the game package so it doesn't
gain access to any internal methods our framework's
game package has access to.
